refactor: fix unit tests file name
